Fix channel download bugs and allow multiple channel URL sources
urllib 'retrieve' will not raise exceptions when receiving a 404, so
add that behavior. This caused empty channel lists when a server was
down, instead of handling this as an error.

Multiple URLs may be provided for channel files, to use this feature
the sources.xml file needs to define channels under a name. Old sources.xml
files remain valid, which means that the "channels" for each source
can either be a URL or a reference to a <channel name...>

Example sources contents used while testing ('ubuntu' is my web server):

<channel name="default">
	<url>http://ubuntu/~dream/zero.gz</url>
	<url>http://ubuntu/~dream/nonexistent.gz</url>
	<url>http://ubuntu/~dream/rytec.channels.xml.gz</url>
</channel>
<source type="gen_xmltv" channels="default">
	<description>Rytec Benelux XMLTV</description>
	<url>http://ubuntu/~dream/zero.gz</url>
	<url>http://unknown.server.bla/bla.gz</url>
	<url>http://ubuntu/~dream/nonexistent.gz</url>
	<url>http://ubuntu/~dream/rytecxmltvbenl.gz</url>
</source>