Fix spurious fails on "unable to open /tmp/epgimport.gz"
The routine removes the file after starting the thread. This introduces a
race condition, if the thread doesn't open the file quickly enough it will
be gone. Fix the race by waiting to remove the file until inside the thread,
which already had the call to remove the temporary file in place.