Small improvements in pixmap handling and code cosmectics (#34)
* Whitespace clean up and other code cosmetics

No functional changes in this commit.

* Load pixmap in the plugin code instead of relying in enigma2

Instead of importing the global variable "selectionpng" from enigma2, load the selection pixmap in the plugin code and use that. This provides safety in case something is changed in the enigma2 code and the "selectionpng" is not available anymore.

* Added icon for non-selected entries in the list

* Load pixmaps from the "icons" folder of the skin

These pixmaps are almost always found in the "icons" folder of every skin and not in the root skin folder.

* Use alpha blending for pixmaps

This works way better for transparent or semi-transparent images and backgrounds.