[xmltvconverter] Fix get_xml_element method:
<programme channel="Akv&#225;rium" start="20211030180000 +0200"
stop="20211030190000 +0200">
       <title lang="cs">Akv&#225;rium</title>
       <desc lang="cs" />
</programme>

When get_xml_method was run with "desc" argument on the element above
it expected that the text was set for the element, which is not the case.

So elements like these were dropped from possible EPG lookup with error message:
[XMLTVConverter] parsing event error: 'NoneType' object has no attribute
'replace',

It is because unescape was run on NoneType object inside get_xml_method,
which raised the exception. To fix this, r variable will be not
reassigned to NoneType but it will stay as '' if text is None
so unescape will not cause crash and element will be considered in
EPG lookup